tests and formatting
